fix: 修复标签管理编辑功能和缓存问题
**Bug 修复**:
- 修复标签管理编辑功能（缓存键解析错误导致所有标签显示相同数据）
- 修复验证码 Base64 编码错误（使用 Java 内置编码器替代自定义实现）
- 修复博客设置页面标题显示问题（动态更新页面标题）

**代码优化**:
- 统一标签字段命名（tagId/tagName → id/name，保持前后端一致）
- 删除 BlogTag 兼容性方法（简化代码结构）
- 优化缓存键解析逻辑（支持包含 # 的表达式解析）
- 将 replace() 改为 replaceAll() 确保所有匹配项都被替换

**功能优化**:
- 优化博客设置管理（删除未使用的配置项，合并标签页）
- 删除未使用的博客配置（主题颜色、创建时间、博主签名等）
- 优化博客设置页面（从 6 个标签页合并为 4 个）
- 同步文章作者字段（使用博客设置中的作者作为默认值）

**测试文件**:
- 更新测试文件以适配新的字段命名
- 修复测试编译错误

+++ ruoyi-framework/src/main/java/com/ruoyi/framework/web/service/SysPasswordService.java
@@ -54,7 +54,7 @@ public class SysPasswordService
             retryCount = 0;
         }
 
-        if (retryCount >= Integer.valueOf(maxRetryCount).intValue())
+        if (retryCount >= maxRetryCount)
         {
             throw new UserPasswordRetryLimitExceedException(maxRetryCount, lockTime);
         }
